% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsdata.R
\docType{data}
\name{blsdata}
\alias{blsdata}
\title{Box Lunch Study - Baseline dataset}
\format{A data frame with 226 rows and 26 variables}
\usage{
data(blsdata)
}
\description{
The variables are as follows:
}
\details{
\itemize{
  \item trt. Treatment
  \item sex. Sex
  \item bmi0. BMI
  \item snackkcal0. Snacking kilo calories
  \item srvgfv0. Serving size of fruits and vegetables
  \item srvgssb0. Serving size of beverages
  \item kcal24h0.
  \item edeq01.
  \item edeq02.
  \item edeq13.
  \item edeq14.
  \item edeq15.
  \item edeq22.
  \item edeq23.
  \item edeq25.
  \item edeq26.
  \item cdrsbody0. Body image
  \item weighfreq0. Weighing frequency
  \item freqff0. Fast food frequency
  \item age. Age
  \item tfactor1.
  \item tfactor2.
  \item tfactor3.
  \item mlhfbias0.
  \item fwahfbias0.
  \item rrvfood. Relative reinforcement of food

}
}
\examples{
data(blsdata)

}
\keyword{datasets}
