% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_node.R
\name{l_node}
\alias{l_node}
\title{Function for determining a pathway}
\usage{
l_node(newtree, node_id = 1, start_criteria = character(0))
}
\arguments{
\item{newtree}{Decision tree generated as a party object}

\item{node_id}{Node ID}

\item{start_criteria}{Character vector}
}
\description{
Decision tree structure
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
