% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_node.R
\name{list_node}
\alias{list_node}
\title{Function for determining a pathway}
\usage{
list_node(newtree, as.party.tree)
}
\arguments{
\item{newtree}{Decision tree generated as a party object}

\item{as.party.tree}{Convert to a party object}
}
\description{
Generates the pathway from the root node to individual terminal nodes of a decision tree generated as a party object using the partykit package.
}
\examples{
list_node(partykit::ctree(Girth ~ ., data = trees), as.party.tree = FALSE)

}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
