% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Function for determining a pathway}
\usage{
makeTransparent(..., alpha)
}
\arguments{
\item{treet}{Decision tree}
}
\description{
Function for determining a pathway
}
\examples{

makeTransparent()

}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
