% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax_mat.R
\name{minmax_mat}
\alias{minmax_mat}
\title{Minmax matrix}
\usage{
minmax_mat(str, varnms, Y, interval)
}
\arguments{
\item{str}{Structure of pathway from the root node in the decision tree to each terminal node}

\item{varnms}{Names of covariates}

\item{Y}{Response variable in the dataset}

\item{interval}{logical. Continuous (interval = FALSE) and Categorical variable (interval = TRUE).}
}
\description{
Identifies splits and relevant criteria
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
\keyword{visualization}
