% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_node.R
\name{path_node}
\alias{path_node}
\title{Function for determining a pathway}
\usage{
path_node(newtree, as.party.tree, idnumber = 0)
}
\arguments{
\item{newtree}{Decision tree generated as a party object}

\item{as.party.tree}{Convert to a party object}

\item{idnumber}{Terminal ID number}
}
\description{
Generates the pathway from the root node to individual terminal nodes of a decision tree generated as a party object using the partykit package.
}
\examples{
path_node(partykit::ctree(Girth ~ ., data = trees), as.party.tree = FALSE)

}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
