% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_minmax.R
\name{plot_minmax}
\alias{plot_minmax}
\title{Plots}
\usage{
plot_minmax(My, X, Y, str, color.type, alpha)
}
\arguments{
\item{My}{Minmax.mat}

\item{X}{Covariates}
}
\description{
Plots
}
\examples{

plot_minmax()

}
\keyword{decision}
\keyword{matrix}
\keyword{pathway}
\keyword{tree}
