% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptree_criteria.R
\name{ptree_criteria}
\alias{ptree_criteria}
\title{Function for determining a pathway}
\usage{
ptree_criteria(newtree, node_id, left)
}
\arguments{
\item{newtree}{decision tree}
}
\description{
Function for determining a pathway
}
\examples{

ptree_criteria()
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
