% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptree_left.R
\name{ptree_left}
\alias{ptree_left}
\title{Function for determining a pathway}
\usage{
ptree_left(newtree, start_id)
}
\arguments{
\item{newtree}{Decision tree}
}
\description{
Function for determining a pathway
}
\examples{

ptree_left()
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
