% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptree_right.R
\name{ptree_right}
\alias{ptree_right}
\title{Function for determining a pathway}
\usage{
ptree_right(newtree, start_id)
}
\arguments{
\item{newtree}{Decision tree}
}
\description{
Function for determining a pathway
}
\examples{

ptree_right()
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
