% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptree_y.R
\name{ptree_y}
\alias{ptree_y}
\title{Function for determining a pathway}
\usage{
ptree_y(newtree, node_id)
}
\arguments{
\item{newtree}{Decision tree}
}
\description{
Function for determining a pathway
}
\examples{

ptree_y()
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
