% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_node.R
\name{l_node}
\alias{l_node}
\title{Function for determining a pathway}
\usage{
l_node(newtree, node_id = 1, start_criteria = character(0))
}
\arguments{
\item{newtree}{Decision tree}
}
\description{
Function for determining a pathway
}
\examples{

l_node()
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
