% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_node.R
\name{l_node}
\alias{l_node}
\title{Function for determining a pathway}
\usage{
l_node(newtree, node_id = 1, start_criteria = character(0), as.party.tree)
}
\arguments{
\item{newtree}{Decision tree generated as a party object}

\item{node_id}{Node ID}

\item{start_criteria}{Character vector}

\item{as.party.tree}{Convert to a party object}
}
\description{
Decision tree structure
}
\examples{
l_node(partykit::ctree(Girth ~ ., data = trees), as.party.tree = FALSE)

}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
