% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Color Scheme}
\usage{
makeTransparent(colortype, alpha)
}
\arguments{
\item{colortype}{Color palette}

\item{alpha}{Transparency}
}
\description{
Function to adjust the transparency and define the color scheme within the visualization.
}
\examples{
makeTransparent(colortype = rainbow(2), alpha = 0.7)

}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
