% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax_mat.R
\name{minmax_mat}
\alias{minmax_mat}
\title{Minmax matrix}
\usage{
minmax_mat(str, varnms, interval)
}
\arguments{
\item{str}{Structure}

\item{varnms}{Variable names}
}
\description{
Minmax matrix
}
\examples{

minmax_mat()

}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
\keyword{visualization}
