% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax_mat.R
\name{minmax_mat}
\alias{minmax_mat}
\title{Minmax matrix}
\usage{
minmax_mat(Y, str, varnms, interval)
}
\arguments{
\item{Y}{Response variable in the dataset}

\item{str}{Decision tree structure}

\item{varnms}{Names of covariates}

\item{interval}{Specify the nature of the response variable Y. Categorical (TRUE) and continuous outcomes (FALSE).}
}
\description{
Identifies splits and relevant criteria
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
\keyword{visualization}
