% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_node.R
\name{path_node}
\alias{path_node}
\title{Function for determining a pathway}
\usage{
path_node(treet)
}
\arguments{
\item{treet}{Decision tree}
}
\description{
Function for determining a pathway
}
\examples{

path_node()
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
