% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptree_criteria.R
\name{ptree_criteria}
\alias{ptree_criteria}
\title{Splitting Criteria}
\usage{
ptree_criteria(newtree, node_id, left)
}
\arguments{
\item{newtree}{Decision tree}

\item{node_id}{Node id}

\item{left}{Splits to the left}
}
\description{
Identifies the splitting criteria for the relevant node leading to lower level inner nodes or a terminal node.
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
