% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptree_left.R
\name{ptree_left}
\alias{ptree_left}
\title{Function for determining a pathway}
\usage{
ptree_left(newtree, start_id)
}
\arguments{
\item{newtree}{Decision tree generated as a party object}

\item{start_id}{Start ID}
}
\description{
Identifies a node that corresponds to the left split
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
