% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptree_right.R
\name{ptree_right}
\alias{ptree_right}
\title{Right Split}
\usage{
ptree_right(newtree, start_id)
}
\arguments{
\item{newtree}{Decision tree generated as a party object}

\item{start_id}{Character vector}
}
\description{
Identifies a node that corresponds to the right split
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
