% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptree_y.R
\name{ptree_y}
\alias{ptree_y}
\title{Function for determining a pathway}
\usage{
ptree_y(newtree, node_id)
}
\arguments{
\item{newtree}{Decision tree generated as a party object}

\item{node_id}{Node ID}
}
\description{
Identifies the predicted outcome value for the relevant node.
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
