% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visTree.R
\name{visTree}
\alias{visTree}
\title{Visualization of subgroups for decision trees}
\usage{
visTree(cond.tree, rng = NULL, interval = FALSE, color.type = 2,
  alpha = 0.2, text.main = 2, text.bar = 2, text.title = 2,
  text.label = 2, text.axis = 2)
}
\arguments{
\item{cond.tree}{Decision tree generated as a party object.}

\item{rng}{Restrict plotting to a particular set of nodes. Default value is set as NULL}

\item{interval}{logical. Continuous (interval = FALSE) and Categorical variable (interval = TRUE).}

\item{color.type}{Color palettes (rainbow_hcl = 1; heat.colors = 2; terrain.colors = 3; topo.colors = 4 ; cm.colors = 5)}

\item{alpha}{Transparency for individual horizonatal colored bars within each subplot. Values between 0 to 1.}

\item{text.main}{Change the size of the main titles}

\item{text.bar}{Change the size of the text in the horizontal bar and below the bar plot}

\item{text.title}{Change the size of the text in the title}

\item{text.label}{Change the size of the axis annotation}

\item{text.axis}{Change the size of the text of axis labels}
}
\description{
This visualization characterizes subgroups defined by a decision tree structure and identifies the range of covariate values associated with outcome values in each subgroup.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
ed<-partykit::extree_data(Ozone ~ ., data = airq)
airct <- partykit::ctree(Ozone ~ ., data = airq)

}
\author{
Ashwini Venkatasubramaniam and Julian Wolfson
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
\keyword{visualization}
