% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visTree.R
\name{visTree}
\alias{visTree}
\title{Visualization}
\usage{
visTree(cond.tree, rng = NULL, interval, as.party.tree = T, color.type,
  alpha)
}
\arguments{
\item{cond.tree}{Decision tree}

\item{rng}{Range}
}
\description{
Visualization
}
\examples{

visTree()

}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
\keyword{visualization}
