% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visTree.R
\name{visTree}
\alias{visTree}
\title{Visualization of subgroups for decision trees}
\usage{
visTree(cond.tree, rng = NULL, interval = FALSE, color.type = 1,
  alpha = 0.5, add.h.axis = TRUE, add.p.axis = TRUE,
  text.round = 1, text.main = 1.5, text.bar = 1.5,
  text.title = 1.5, text.label = 1.5, text.axis = 1.5,
  text.percentile = 0.7, density.line = TRUE)
}
\arguments{
\item{cond.tree}{Decision tree generated as a party object.}

\item{rng}{Restrict plotting to a particular set of nodes. Default value is set as NULL.}

\item{interval}{logical. Continuous outcome (interval = FALSE) and Categorical outcome (interval = TRUE).}

\item{color.type}{Color palettes (rainbow_hcl = 1; heat_hcl = 2; terrain_hcl = 3; sequential_hcl = 4 ; diverge_hcl = 5)}

\item{alpha}{Transparency for horizontal colored bars in each subplot. Values between 0 to 1.}

\item{add.h.axis}{logical. Add axis for the outcome distribution (add.h.axis = TRUE), remove axis for the outcome (add.h.axis = FALSE).}

\item{add.p.axis}{logical. Add axis for the percentiles (add.p.axis = TRUE) computed over covariate values, remove axis for the percentiles (add.p.axis = FALSE).}

\item{text.round}{Round the threshold displayed on the horizontal bar}

\item{text.main}{Change the size of the main titles}

\item{text.bar}{Change the size of the text in the horizontal bar}

\item{text.title}{Change the size of the text in the title}

\item{text.label}{Change the size of the axis annotation}

\item{text.axis}{Change the size of the text of axis labels}

\item{text.percentile}{Change the size of the percentile title}

\item{density.line}{logical. Draw a density line. (density.line = TRUE).}
}
\description{
This visualization characterizes subgroups defined by a decision tree structure and identifies the range of covariate values associated with outcome values in each subgroup.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
ed<-partykit::extree_data(Ozone ~ ., data = airq)
airct <- partykit::ctree(Ozone ~ ., data = airq)
visTree(airct, text.bar = 1.1, text.percentile = 0.8)
}
\author{
Ashwini Venkatasubramaniam and Julian Wolfson
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
\keyword{visualization}
