% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visTree.R
\name{visTree}
\alias{visTree}
\title{Visualization of subgroups for decision trees}
\usage{
visTree(cond.tree, rng = NULL, interval, as.party.tree, color.type = 2,
  alpha = 0.4)
}
\arguments{
\item{cond.tree}{Decision tree generated as a party object.}

\item{rng}{Restrict plotting to a particular set of nodes. Default value is set as NULL}

\item{interval}{logical. Continuous (interval = FALSE) and Categorical variable (interval = TRUE).}

\item{as.party.tree}{logical. Convert to a party object from other decision tree types such as an rpart object.}

\item{color.type}{Color palettes (rainbow = 1; heat.colors = 2; terrain.colors = 3; topo.colors = 4 ; cm.colors = 5)}

\item{alpha}{Transparency for individual horizonatal colored bars within each subplot. Values between 0 to 1.}
}
\description{
This visualization characterizes subgroups defined by a decision tree structure and identifies the range of covariate values associated with outcome values in each subgroup.
}
\examples{
visTree(partykit::ctree(Girth ~ ., data = trees), interval = FALSE, as.party.tree = FALSE)

}
\author{
Ashwini Venkatasubramaniam and Julian Wolfson
}
\keyword{decision}
\keyword{pathway}
\keyword{tree}
\keyword{visualization}
